CLEANUP: Prolog flag handling.
Cleanup the implementation, mostly by splitting too large functions.
Clarified documentation with respect to thread interaction.  Added
an option `local(true)` to create_prolog_flag/3 which create a flag
only for the calling thread.  Clarified and fixed default behaviour,
which is for create_prolog_flag/3 to set the flag _globally_, i.e.,
make it available with this value for any thread.  This now updates
_both_ the local and global flag table (before it updated only the
global table).

+++ src/pl-fli.c
@@ -5495,7 +5495,7 @@ PL_set_prolog_flag(const char *name, int type, ...)
 { GET_LD
   va_list args;
   int rval = true;
-  unsigned short flags = ((unsigned short)type & FF_MASK);
+  unsigned int flags = ((unsigned int)type & FF_MASK);
   fid_t fid;
   term_t av;
